requirements.txt
django
gunicorn
openpyxl>=3.1.2
python-dotenv
psycopg2

whitenoise
djangorestframework